+++ .vscode/settings.json
@@ -7,5 +7,8 @@
         "--doctest-modules"
     ],
     "python.testing.unittestEnabled": false,
-    "python.testing.pytestEnabled": true
+    "python.testing.pytestEnabled": true,
+    "python.analysis.extraPaths": [
+        "./easy2homeassistant"
+    ]
 }
