+++ easy2homeassistant/logging_config.py
@@ -16,7 +16,7 @@ LOGGER_FORMAT = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
 LOG_FILE_PATH = "easy2homeassistant.log"
 
 
-def get_logger(name: str):
+def configure_logging(logging_level="INFO"):
     """Return a logger with the specified name."""
     console_handler = colorlog.StreamHandler(sys.stdout)
     console_formatter = colorlog.ColoredFormatter(
@@ -28,8 +28,9 @@ def get_logger(name: str):
     file_formatter = logging.Formatter(LOGGER_FORMAT)
     file_handler.setFormatter(file_formatter)
 
-    logger = logging.getLogger(name)
-    logger.addHandler(console_handler)
-    logger.addHandler(file_handler)
+    logging.basicConfig(handlers=[console_handler, file_handler])
 
-    return logger
+
+def set_logging_level(logging_level):
+    """Set the logging level for the root logger."""
+    logging.root.setLevel(logging_level)
